day23-05
micro + gin 服务构建 http 加 rpc服务的调用

+++ day23/goods_service/models/goods.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Goods struct {
 	gorm.Model
-	ShopId     int    `json:"shop_id"`
-	CategoryId int    `json:"category_id"`
+	ShopId     int64  `json:"shop_id"`
+	CategoryId int64  `json:"category_id"`
 	GoodName   string `json:"good_name"`
 }
